Moved some classes out of Faction.cs that don't really belong in  this file.
