+++ Pulsar4X/Pulsar4X.Lib/Entities/Faction.cs
@@ -16,599 +16,6 @@ using log4net;
 
 namespace Pulsar4X.Entities
 {
-    public class FactionContact
-    {
-        /// <summary>
-        /// The detected ship.
-        /// </summary>
-        public ShipTN ship { get; set; }
-
-        /// <summary>
-        /// Or it could be a detected missile.
-        /// </summary>
-        public OrdnanceGroupTN missileGroup { get; set; }
-
-        /// <summary>
-        /// Detected via thermal.
-        /// </summary>
-        public bool thermal { get; set; }
-
-        /// <summary>
-        /// Tick thermal detect event happened on.
-        /// </summary>
-        public uint thermalTick { get; set; }
-
-        /// <summary>
-        /// Detected via EM.
-        /// </summary>
-        public bool EM { get; set; }
-
-        /// <summary>
-        /// EM signature
-        /// </summary>
-        public int EMSignature { get; set; }
-
-        /// <summary>
-        /// Tick detected via EM.
-        /// </summary>
-        public uint EMTick { get; set; }
-
-        /// <summary>
-        /// Detected via Actives.
-        /// </summary>
-        public bool active { get; set; }
-
-        /// <summary>
-        /// Tick active detection event occurred on.
-        /// </summary>
-        public uint activeTick { get; set; }
-
-        /// <summary>
-        /// Initializer for detected ship event. FactionContact is the detector side of what is detected, while ShipTN itself stores the detectee side. 
-        /// multiple of these can exist, but only 1 per faction hopefully.
-        /// </summary>
-        /// <param name="DetectedShip">Ship detected.</param>
-        /// <param name="Thermal">Was the detection thermal based?</param>
-        /// <param name="em">Detection via EM?</param>
-        /// <param name="Active">Active detection?</param>
-        /// <param name="tick">What tick did this detection event occur on?</param>
-        public FactionContact(Faction CurrentFaction, ShipTN DetectedShip, bool Thermal, bool em, int EMSig, bool Active, uint tick)
-        {
-            ship = DetectedShip;
-            missileGroup = null;
-            thermal = Thermal;
-            EM = em;
-            EMSignature = EMSig;
-            active = Active;
-
-            String Contact = "New contact detected:";
-
-            if (thermal == true)
-            {
-                thermalTick = tick;
-                Contact = String.Format("{0} Thermal Signature {1}", Contact, DetectedShip.CurrentThermalSignature);
-            }
-
-            if (EM == true)
-            {
-                EMTick = tick;
-                Contact = String.Format("{0} EM Signature {1}", Contact, EMSignature);
-            }
-
-            if (active == true)
-            {
-                activeTick = tick;
-                Contact = String.Format("{0} TCS {1}", Contact, DetectedShip.TotalCrossSection);
-            }
-
-            /// <summary>
-            /// Print to the message log.
-            /// </summary>
-            MessageEntry NMsg = new MessageEntry(MessageEntry.MessageType.ContactNew, DetectedShip.ShipsTaskGroup.Contact.Position.System, DetectedShip.ShipsTaskGroup.Contact,
-                                                 GameState.Instance.GameDateTime, GameState.Instance.LastTimestep, Contact);
-
-            CurrentFaction.MessageLog.Add(NMsg);
-
-            /// <summary>
-            /// Inform SimEntity.
-            /// </summary>
-            GameState.SE.SetInterrupt(InterruptType.NewSensorContact);
-
-        }
-
-        /// <summary>
-        /// Initializer for detected missile event. FactionContact is the detector side of what is detected, while OrdnanceTN itself stores the detectee side. 
-        /// multiple of these can exist, but only 1 per faction hopefully.
-        /// </summary>
-        /// <param name="DetectedMissile">Missile detected.</param>
-        /// <param name="Thermal">Was the detection thermal based?</param>
-        /// <param name="em">Detection via EM?</param>
-        /// <param name="Active">Active detection?</param>
-        /// <param name="tick">What tick did this detection event occur on?</param>
-        public FactionContact(Faction CurrentFaction, OrdnanceGroupTN DetectedMissileGroup, bool Thermal, bool em, bool Active, uint tick)
-        {
-            missileGroup = DetectedMissileGroup;
-            ship = null;
-            thermal = Thermal;
-            EM = em;
-            active = Active;
-            MessageEntry NMsg;
-
-            String Contact = "New contact detected:";
-
-            if (thermal == true)
-            {
-                thermalTick = tick;
-                Contact = String.Format("{0} Thermal Signature {1} x{2}", Contact, (int)Math.Ceiling(DetectedMissileGroup.missiles[0].missileDef.totalThermalSignature), DetectedMissileGroup.missiles.Count);
-            }
-
-            if (EM == true)
-            {
-                EMTick = tick;
-                if (DetectedMissileGroup.missiles[0].missileDef.aSD != null)
-                {
-                    Contact = String.Format("{0} EM Signature {1} x{2}", Contact, DetectedMissileGroup.missiles[0].missileDef.aSD.gps, DetectedMissileGroup.missiles.Count);
-                }
-                else
-                {
-                    Contact = String.Format("Error with {0} : has EM signature but no active sensor.", Contact);
-                    NMsg = new MessageEntry(MessageEntry.MessageType.Error, DetectedMissileGroup.contact.Position.System, DetectedMissileGroup.contact,
-                                                 GameState.Instance.GameDateTime, GameState.Instance.LastTimestep, Contact);
-
-                    CurrentFaction.MessageLog.Add(NMsg);
-                }
-
-            }
-
-            if (active == true)
-            {
-                activeTick = tick;
-                Contact = String.Format("{0} TCS {1} x{2}", Contact, (int)Math.Ceiling(DetectedMissileGroup.missiles[0].missileDef.size), DetectedMissileGroup.missiles.Count);
-            }
-
-            /// <summary>
-            /// print to the message log.
-            /// </summary>
-            NMsg = new MessageEntry(MessageEntry.MessageType.ContactNew, DetectedMissileGroup.contact.Position.System, DetectedMissileGroup.contact,
-                                                 GameState.Instance.GameDateTime, GameState.Instance.LastTimestep, Contact);
-
-            CurrentFaction.MessageLog.Add(NMsg);
-
-            /// <summary>
-            /// Inform SimEntity.
-            /// </summary>
-            GameState.SE.SetInterrupt(InterruptType.NewSensorContact);
-        }
-
-
-
-        /// <summary>
-        /// Each tick every faction contact should be updated on the basis of collected sensor data.
-        /// </summary>
-        /// <param name="Thermal">Thermal detection?</param>
-        /// <param name="Em">Detected on EM?</param>
-        /// <param name="Active">Detected by actives?</param>
-        /// <param name="tick">Current tick.</param>
-        public void updateFactionContact(Faction CurrentFaction, bool Thermal, bool Em, int EMSig, bool Active, uint tick)
-        {
-            if (thermal == Thermal && EM == Em && active == Active)
-            {
-                return;
-            }
-
-            String Contact = "N/A";
-            MessageEntry.MessageType type = MessageEntry.MessageType.Count;
-
-            if (Thermal == false && Em == false && Active == false)
-            {
-                Contact = "Existing contact lost";
-                type = MessageEntry.MessageType.ContactLost;
-            }
-            else
-            {
-                Contact = "Update on existing contact:";
-                type = MessageEntry.MessageType.ContactUpdate;
-
-                if (thermal == false && Thermal == true)
-                {
-                    /// <summary>
-                    /// New thermal detection event, message logic should be here.
-                    /// </summary>
-                    thermalTick = tick;
-
-                    if (ship != null)
-                        Contact = String.Format("{0} Thermal Signature {1}", Contact, ship.CurrentThermalSignature);
-                    else if (missileGroup != null)
-                        Contact = String.Format("{0} Thermal Signature {1} x{2}", Contact, (int)Math.Ceiling(missileGroup.missiles[0].missileDef.totalThermalSignature), missileGroup.missiles.Count);
-                    else
-                    {
-                        type = MessageEntry.MessageType.Error;
-                        Contact = "Error: Both ship and missile are null in UpdateFactionContact.";
-                    }
-                }
-                else if (thermal == true && Thermal == false)
-                {
-                    /// <summary>
-                    /// Thermal contact lost.
-                    /// </summary>
-                    Contact = String.Format("{0} Thermal contact lost", Contact);
-                }
-
-                if (EM == false && Em == true)
-                {
-                    /// <summary>
-                    /// New EM detection event, message logic should be here.
-                    /// </summary>
-                    EMTick = tick;
-
-                    EMSignature = EMSig;
-
-
-                    if (ship != null)
-                        Contact = String.Format("{0} EM Signature {1}", Contact, EMSignature);
-                    else if (missileGroup != null)
-                    {
-                        if (missileGroup.missiles[0].missileDef.aSD != null)
-                        {
-                            Contact = String.Format("{0} EM Signature {1} x{2}", Contact, missileGroup.missiles[0].missileDef.aSD.gps, missileGroup.missiles.Count);
-                        }
-                        else
-                        {
-                            type = MessageEntry.MessageType.Error;
-                            Contact = "Error: Missile lacks a sensor, but is emitting EM in UpdateFactionContact.";
-                        }
-                    }
-                    else
-                    {
-                        type = MessageEntry.MessageType.Error;
-                        Contact = "Error: Both ship and missile are null in UpdateFactionContact.";
-                    }
-                }
-                if (EM == true && Em == false)
-                {
-                    EMSignature = -1;
-                    /// <summary>
-                    /// EM contact lost.
-                    /// </summary>
-                    Contact = String.Format("{0} EM contact lost", Contact);
-                }
-
-                if (active == false && Active == true)
-                {
-                    /// <summary>
-                    /// New active detection event, message logic should be here.
-                    /// </summary>
-                    activeTick = tick;
-
-
-
-                    if (ship != null)
-                        Contact = String.Format("{0} TCS {1}", Contact, ship.TotalCrossSection);
-                    else if (missileGroup != null)
-                        Contact = String.Format("{0} TCS_MSP {1} x{2}", Contact, (int)Math.Ceiling(missileGroup.missiles[0].missileDef.size), missileGroup.missiles.Count);
-                    else
-                    {
-                        type = MessageEntry.MessageType.Error;
-                        Contact = "Error: Both ship and missile are null in UpdateFactionContact.";
-                    }
-                }
-                if (active == true && Active == false)
-                {
-                    /// <summary>
-                    /// Active contact lost.
-                    /// </summary>
-
-                    Contact = String.Format("{0} Active contact lost", Contact);
-                }
-
-            }
-
-            SystemContact SysCon = null;
-
-            if (ship == null)
-                SysCon = missileGroup.contact;
-            else if (missileGroup == null)
-                SysCon = ship.ShipsTaskGroup.Contact;
-
-            MessageEntry NMsg = new MessageEntry(type, SysCon.Position.System, SysCon,
-                                                 GameState.Instance.GameDateTime, GameState.Instance.LastTimestep, Contact);
-
-            CurrentFaction.MessageLog.Add(NMsg);
-
-            thermal = Thermal;
-            EM = Em;
-            active = Active;
-        }
-    }
-
-    /// <summary>
-    /// I need a new class to store a detected contacts list for every system.
-    /// </summary>
-    public class DetectedContactsList
-    {
-        /// <summary>
-        /// This was the original faction detected contacts list, now moved to this class container so that I can put this in a dictionary with System.
-        /// </summary>
-        public Dictionary<ShipTN, FactionContact> DetectedContacts { get; set; }
-
-        /// <summary>
-        /// A DetectionEntity subclass would have come in handy perhaps. This is the detected contact list that is specific to missiles.
-        /// </summary>
-        public Dictionary<OrdnanceGroupTN, FactionContact> DetectedMissileContacts { get; set; }
-
-        /// <summary>
-        /// Constructor for this new list.
-        /// </summary>
-        public DetectedContactsList()
-        {
-            DetectedContacts = new Dictionary<ShipTN, FactionContact>();
-            DetectedMissileContacts = new Dictionary<OrdnanceGroupTN, FactionContact>();
-        }
-    }
-
-    public class FactionSystemDetection
-    {
-        /// <summary>
-        /// Faction that owns this list.
-        /// </summary>
-        public Faction Faction { get; set; }
-
-        /// <summary>
-        /// The system these contacts are stored for.
-        /// </summary>
-        public StarSystem System { get; set; }
-
-        /// <summary>
-        /// Last time this contact index was spotted via thermals.
-        /// </summary>
-        public BindingList<int> Thermal { get; set; }
-
-        /// <summary>
-        /// Last time this contact index was spotted via EM.
-        /// </summary>
-        public BindingList<int> EM { get; set; }
-
-        /// <summary>
-        /// Last time this contact index was spotted via Active.
-        /// </summary>
-        public BindingList<int> Active { get; set; }
-
-        /// <summary>
-        /// Creates a faction contact list for the specified system.
-        /// </summary>
-        /// <param name="system">System indicated.</param>
-        public FactionSystemDetection(Faction Fact, StarSystem system)
-        {
-            Faction = Fact;
-            System = system;
-
-            Thermal = new BindingList<int>();
-            EM = new BindingList<int>();
-            Active = new BindingList<int>();
-
-            Thermal.RaiseListChangedEvents = false;
-            EM.RaiseListChangedEvents = false;
-            Active.RaiseListChangedEvents = false;
-
-            for (int loop = 0; loop < system.SystemContactList.Count; loop++)
-            {
-                Thermal.Add(0);
-                EM.Add(0);
-                Active.Add(0);
-            }
-
-            Thermal.RaiseListChangedEvents = true;
-            EM.RaiseListChangedEvents = true;
-            Active.RaiseListChangedEvents = true;
-
-            Thermal.ResetBindings();
-            EM.ResetBindings();
-            Active.ResetBindings();
-        }
-
-        /// <summary>
-        /// pushes a contact onto the end of the list since that is where all new contacts will be added.
-        /// </summary>
-        public void AddContact()
-        {
-            Thermal.Add(0);
-            EM.Add(0);
-            Active.Add(0);
-        }
-
-        /// <summary>
-        /// Removes a contact at the specified index.
-        /// </summary>
-        /// <param name="RemIndex">index to be removed.</param>
-        public void RemoveContact(int RemIndex)
-        {
-            Thermal.RemoveAt(RemIndex);
-            EM.RemoveAt(RemIndex);
-            Active.RemoveAt(RemIndex);
-        }
-    }
-
-    public class MessageEntry
-    {
-        /// <summary>
-        /// Message types that will be printed to the event log.
-        /// </summary>
-        public enum MessageType
-        {
-            ColonyLacksCI,
-            ColonyLacksMinerals,
-
-            ContactNew,
-            ContactUpdate,
-            ContactLost,
-
-            Firing,
-            FiringHit,
-            FiringMissed,
-            FiringNoAvailableOrdnance,
-            FiringNoLoadedOrdnance,
-            FiringRecharging,
-            FiringZeroHitChance,
-
-            LaunchTubeReloaded,
-            LaunchTubeNoOrdnanceToReload,
-
-            MissileHit,
-            MissileIntercepted,
-            MissileInterceptEvent,
-            MissileLostFireControl,
-            MissileLostTracking,
-            MissileMissed,
-            MissileOutOfFuel,
-
-            OrdersCompleted,
-            OrdersNotCompleted,
-
-            PotentialFleetInterception,
-
-            ShieldRecharge,
-
-            ShipDamage,
-            ShipDamageReport,
-
-            Error,
-            Count
-        }
-
-#if LOG4NET_ENABLED
-        /// <summary>
-        /// Ship Logger:
-        /// </summary>
-        public static readonly ILog logger = LogManager.GetLogger(typeof(MessageEntry));
-#endif
-
-        /// <summary>
-        /// What specific type of message is this?
-        /// </summary>
-        public MessageType TypeOf { get; set; }
-
-        /// <summary>
-        /// which starsystem does this message occur in.
-        /// </summary>
-        public StarSystem Location { get; set; }
-
-        /// <summary>
-        /// Which tg/planet/pop are we referencing?
-        /// </summary>
-        public StarSystemEntity entity { get; set; }
-
-        /// <summary>
-        /// When was this message sent?
-        /// </summary>
-        public DateTime TimeOfMessage { get; set; }
-
-        /// <summary>
-        /// How long since the last time increment?
-        /// </summary>
-        public int TimeSlice { get; set; }
-        /// <summary>
-        /// Text of the message for the log.
-        /// </summary>
-        public String Text { get; set; }
-
-
-        /// <summary>
-        /// MessageEntry constructs a specific message with the relevant location, time, and entity(if applicable). 
-        /// </summary>
-        /// <param name="Loc">Starsystem message is from.</param>
-        /// <param name="Ref">Starsystementity the message refers to.</param>
-        /// <param name="Time">Game time of message.</param>
-        /// <param name="timeSlice">Time since last increment.</param>
-        /// <param name="text">text of the message.</param>
-        public MessageEntry(MessageType Type, StarSystem Loc, StarSystemEntity Ref, DateTime Time, int timeSlice, string text)
-        {
-            TypeOf = Type;
-            Location = Loc;
-            entity = Ref;
-            TimeOfMessage = Time;
-            TimeSlice = timeSlice;
-            Text = text;
-
-
-#if LOG4NET_ENABLED
-            if (TypeOf == MessageType.Error)
-            {
-                String Entry = String.Format("Faction Message Logging Error: Type:{0} | Time:{1} | Location:{2} - TimeSlice:{3}: Text:{4}\n", TypeOf, TimeOfMessage, Location, TimeSlice, Text);
-                logger.Debug(Entry);
-            }
-#endif
-        }
-    }
-
-    /// <summary>
-    /// This is intended to be part of a larger dictionary that separates these by system. Component and Ship data is stored for every Point defense enabled FC here.
-    /// </summary>
-    public class PointDefenseList
-    {
-        /// <summary>
-        /// The component and which ship it is on.
-        /// </summary>
-        public Dictionary<ComponentTN, ShipTN> PointDefenseFC { get; set; }
-
-        /// <summary>
-        /// FCType of the base component, again, no pointer exists to the component def so this must be stored.
-        /// as always, false = MFC, true = BFC.
-        /// </summary>
-        public Dictionary<ComponentTN, bool> PointDefenseType { get; set; }
-
-
-        /// <summary>
-        /// Constructor for PDList.
-        /// </summary>
-        public PointDefenseList()
-        {
-            PointDefenseFC = new Dictionary<ComponentTN, ShipTN>();
-            PointDefenseType = new Dictionary<ComponentTN, bool>();
-        }
-
-
-#warning When jump transits are fully implemented, PointDefenseFC listings will have to be moved as appropriate, be sure to handle that.
-
-        /// <summary>
-        /// Handles adding a new FC to the list.
-        /// </summary>
-        /// <param name="Comp">Fire control component to add</param>
-        /// <param name="Ship">Ship the FC is on.</param>
-        /// <param name="Type">Type of FC.</param>
-        public void AddComponent(ComponentTN Comp, ShipTN Ship, bool Type)
-        {
-            if (PointDefenseFC.ContainsKey(Comp) == false)
-            {
-                PointDefenseFC.Add(Comp, Ship);
-            }
-
-            if (PointDefenseType.ContainsKey(Comp) == false)
-            {
-                PointDefenseType.Add(Comp, Type);
-            }
-        }
-
-        /// <summary>
-        /// Handles removing an existing FC from the list.
-        /// </summary>
-        /// <param name="Comp">Fire control component to remove</param>
-        /// <param name="Ship">Ship the FC is on.</param>
-        /// <param name="Type">Type of FC.</param>
-        public void RemoveComponent(ComponentTN Comp)
-        {
-            if (PointDefenseFC.ContainsKey(Comp) == true)
-            {
-                PointDefenseFC.Remove(Comp);
-            }
-
-            if (PointDefenseType.ContainsKey(Comp) == true)
-            {
-                PointDefenseType.Remove(Comp);
-            }
-        }
-    }
-
-
     public class Faction : GameEntity
     {
         /// <summary>
